% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sofmldsc_gc.R
\name{sofmldsc_gc}
\alias{sofmldsc_gc}
\title{用于计算两个性状之间的遗传度以及遗传关联性}
\usage{
sofmldsc_gc(keyssh, GWAS1summary, GWAS2summary, GWAS1name, GWAS2name, pop)
}
\arguments{
\item{keyssh}{学号}

\item{GWAS1summary}{预处理好的GWAS summay1}

\item{GWAS2summary}{预处理好的GWAS summay2}

\item{GWAS1name}{GWAS summay1名称}

\item{GWAS2name}{GWAS summay2名称}

\item{pop}{参考人群EUR欧洲 EAS亚洲}
}
\description{
用于计算两个性状之间的遗传度以及遗传关联性
}
