% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sofmldsc_h2.R
\name{sofmldsc_h2}
\alias{sofmldsc_h2}
\title{用于计算两个性状之间的遗传度以及遗传关联性}
\usage{
sofmldsc_h2(keyssh, GWAS1summary, GWAS1name, pop)
}
\arguments{
\item{keyssh}{学号}

\item{GWAS1summary}{预处理好的GWAS summay}

\item{GWAS1name}{GWAS summay名称}

\item{pop}{参考人群EUR欧洲 EAS亚洲}
}
\description{
用于计算两个性状之间的遗传度以及遗传关联性
}
